subtitle-api/requirements.txt
#逐字稿
fpdf
yt-dlp
pytube
fastapi
uvicorn
reportlab
youtube-transcript-api

# PDF 產生
reportlab 
requests
pillow
fpdf2